+++ ogc-api-features-json-ios/OAFTemporal.h
@@ -24,11 +24,13 @@ extern NSString * const OAF_TRS;
 @interface OAFTemporal : OAFFeaturesObject
 
 /**
- * Begin and end times of the time interval. The timestamps are in the
- * coordinate reference system specified in `trs`. By default this is the
- * Gregorian calendar.
+ * One or more time intervals that describe the temporal extent of the
+ * dataset. The value `null` is supported and indicates an unbounded
+ * interval end. In the Core only a single time interval is supported.
+ * Extensions may support multiple intervals. If multiple intervals are
+ * provided, the union of the intervals describes the temporal extent.
  */
-@property (nonatomic, strong) NSMutableArray<NSString *> *interval;
+@property (nonatomic, strong) NSMutableArray<NSMutableArray<NSString *> *> *interval;
 
 /**
  * Coordinate reference system of the coordinates in the temporal extent
@@ -54,5 +56,36 @@ extern NSString * const OAF_TRS;
  *  @return new temporal
  */
 -(instancetype) initWithTree: (NSDictionary *) tree;
+
+/**
+ * Get the interval collection count
+ *
+ * @return count
+ */
+-(int) intervalCount;
+
+/**
+ * Get the first interval
+ *
+ * @return interval
+ */
+-(NSMutableArray<NSString *> *) firstInterval;
+
+/**
+ * Get the interval at the index
+ *
+ * @param index
+ *            0 based index
+ * @return interval
+ */
+-(NSMutableArray<NSString *> *) intervalAtIndex: (int) index;
+
+/**
+ * Add an interval
+ *
+ * @param interval
+ *            single interval
+ */
+-(void) addInterval: (NSMutableArray<NSString *> *) interval;
 
 @end